Expand grafana display for separate channels of Tesla inveter
This version adds two graphs; one is a stacked graph of the power
output of the four MPPT channels; the second shows both the voltage
and the current of each of the four channels.

Two additional widgets show the status of the inverter as a whole and
of each channel.

Changes are also included here to fit some differences in the data
provided by the GW2: the voltage comes from battery rather than site
and is 240V rather than 120V.